restructured frontend
